requirements.txt
reflex
langchain
langgraph
langchain_community
ollama
sqlalchemy
python-dotenv
tavily-python
pymupdf
faiss-cpu
python-docx
Pillow
pytesseract
openpyxl
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
SpeechRecognition
PyAudio
langchain-ollama